requirements.txt
alembic==1.9.4
altgraph==0.17.3
blinker==1.5
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
dominate==2.7.0
Flask==2.2.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
flaskwebgui==1.0.3
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mysql-connector==2.2.9
mysql-connector-python==8.0.32
mysql-connector-python-rf==2.2.2
mysqlclient==2.1.1
numpy==1.24.1
pandas==1.5.2
pefile==2023.2.7
protobuf==3.20.3
psutil==5.9.4
psycopg2==2.9.5
PyFladesk==1.1
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
pywin32-ctypes==0.2.0
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.4
typing_extensions==4.5.0
urllib3==1.26.14
visitor==0.1.3
waitress==2.1.2
Werkzeug==2.2.3
WTForms==3.0.1
WTForms-SQLAlchemy==0.3
zipp==3.15.0
